+++ json_annotation/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix a potential error with `checked: true` when `ArgumentError.message` is 
   `null`.
 - Updated `JsonSerializable.fromJson` to handle `null` values.
+- Deprecate `JsonSerializable` `defaults` and `withDefaults()`.
 
 ## 4.0.0